Added swap trigger and handler
