+++ plugin/copytrader/const.go
@@ -1,0 +1,6 @@
+package copytrader
+
+const (
+	UniswapV2RouterAddress   = "0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D"
+	SwapExactTokensForTokens = "38ed1739"
+)
