+++ plugin/copytrader/transaction.go
@@ -1,0 +1,97 @@
+package copytrader
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"time"
+
+	gcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/hibiken/asynq"
+	"github.com/sirupsen/logrus"
+	"github.com/vultisig/mobile-tss-lib/tss"
+	vcommon "github.com/vultisig/verifier/common"
+	vtypes "github.com/vultisig/verifier/types"
+	"github.com/vultisig/vultiserver/contexthelper"
+)
+
+func (p *Plugin) HandleSwapTask(c context.Context, t *asynq.Task) error {
+	ctx, cancel := context.WithTimeout(c, 5*time.Minute)
+	defer cancel()
+
+	if err := contexthelper.CheckCancellation(ctx); err != nil {
+		p.logger.WithError(err).Warn("Context cancelled, skipping trigger")
+		return err
+	}
+	var swapTask SwapTask
+	if err := json.Unmarshal(t.Payload(), &swapTask); err != nil {
+		p.logger.WithError(err).Error("Failed to unmarshal swapTask payload")
+		return fmt.Errorf("failed to unmarshal swapTask payload: %s, %w", err, asynq.SkipRetry)
+	}
+
+	//TODO: implement aim <-> policy db
+	//TODO: trigger swaps
+	return nil
+}
+
+func (p *Plugin) ProposeTransactions(policy vtypes.PluginPolicy) ([]vtypes.PluginKeysignRequest, error) {
+	//TODO implement me
+	panic("implement me")
+}
+
+func (p *Plugin) initSign(
+	ctx context.Context,
+	req vtypes.PluginKeysignRequest,
+	pluginPolicy vtypes.PluginPolicy,
+) error {
+	sigs, err := p.signer.Sign(ctx, req)
+	if err != nil {
+		p.logger.WithError(err).Error("Keysign failed")
+		return fmt.Errorf("failed to sign transaction: %w", err)
+	}
+
+	if len(sigs) != 1 {
+		p.logger.
+			WithField("sigs_count", len(sigs)).
+			Error("expected only 1 message+sig per request for evm")
+		return fmt.Errorf("failed to sign transaction: invalid signature count: %d", len(sigs))
+	}
+	var sig tss.KeysignResponse
+	for _, s := range sigs {
+		sig = s
+	}
+
+	err = p.SigningComplete(ctx, sig, req, pluginPolicy)
+	if err != nil {
+		p.logger.WithError(err).Error("failed to complete signing process (broadcast tx)")
+		return fmt.Errorf("failed to complete signing process: %w", err)
+	}
+	return nil
+}
+
+func (p *Plugin) SigningComplete(
+	ctx context.Context,
+	signature tss.KeysignResponse,
+	signRequest vtypes.PluginKeysignRequest,
+	_ vtypes.PluginPolicy,
+) error {
+	tx, err := p.eth.Send(
+		ctx,
+		gcommon.FromHex(signRequest.Transaction),
+		gcommon.Hex2Bytes(signature.R),
+		gcommon.Hex2Bytes(signature.S),
+		gcommon.Hex2Bytes(signature.RecoveryID),
+	)
+	if err != nil {
+		p.logger.WithError(err).WithField("tx_hex", signRequest.Transaction).Error("p.eth.Send")
+		return fmt.Errorf("p.eth.Send(tx_hex=%s): %w", signRequest.Transaction, err)
+	}
+
+	p.logger.WithFields(logrus.Fields{
+		"from_public_key": signRequest.PublicKey,
+		"to_address":      tx.To().Hex(),
+		"hash":            tx.Hash().Hex(),
+		"chain":           vcommon.Ethereum.String(),
+	}).Info("tx successfully signed and broadcasted")
+	return nil
+}
