+++ plugin/copytrader/watcher.go
@@ -2,14 +2,24 @@ package copytrader
 
 import (
 	"context"
+	"encoding/hex"
+	"encoding/json"
+	"fmt"
 	"math/big"
 	"time"
 
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/sirupsen/logrus"
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/vultisig/recipes/sdk/evm/codegen/uniswapv2_router"
 )
 
 func (p *Plugin) WatchSwap(ctx context.Context) {
+	var uniswapABI abi.ABI
+	err := json.Unmarshal([]byte(uniswapv2_router.Uniswapv2RouterMetaData.ABI), &uniswapABI)
+	if err != nil {
+		panic(err)
+	}
+
 	for {
 		select {
 		case <-ctx.Done():
@@ -31,31 +41,52 @@ func (p *Plugin) WatchSwap(ctx context.Context) {
 					continue
 				}
 
+				// Process txs to find UniswapV2Router interactions
 				for _, tx := range block.Transactions() {
+					if tx.To() == nil {
+						continue
+					}
 					// is Uniswap tx check
 					if tx.To().String() == UniswapV2RouterAddress {
-						txReceipt, err := p.ethRpc.TransactionReceipt(ctx, tx.Hash())
-						if err != nil {
-							p.logger.WithError(err).Error("failed to get block")
+						inputBytes := tx.Data()
+						signature, data := inputBytes[:4], inputBytes[4:]
+						if hex.EncodeToString(signature) != SwapExactTokensForTokens {
 							continue
 						}
 
-						for _, log := range txReceipt.Logs {
-							if log.Topics[0] == UniswapSwapTopic {
-								signer := types.LatestSignerForChainID(tx.ChainId())
-								sender, err := signer.Sender(tx)
-								if err != nil {
-									p.logger.WithError(err).Error("failed to get signer")
-									continue
-								}
-								p.logger.WithFields(logrus.Fields{
-									"sender": sender.String(),
-									"txHash": tx.Hash().String(),
-									"pair":   log.Address.String(),
-								})
-								//TODO: Trigger swaps there
-							}
+						method, err := uniswapABI.MethodById(signature)
+						if err != nil {
+							p.logger.WithError(err).Error("unknown method")
+							continue
 						}
+
+						// Getting args from tx to find necessary info
+						var args = make(map[string]interface{})
+						err = method.Inputs.UnpackIntoMap(args, data)
+						if err != nil {
+							p.logger.WithError(err).Error("failed to unpack data")
+							continue
+						}
+
+						path := args["path"]
+						tokens, valid := path.([]common.Address)
+						if !valid {
+							p.logger.Error("invalid path", path)
+							continue
+						}
+
+						amountIn, _ := new(big.Int).SetString(fmt.Sprint(args["amountIn"]), 10)
+						to := args["to"]
+						sender, valid := to.(common.Address)
+						if !valid {
+							p.logger.Error("invalid sender", to)
+							continue
+						}
+
+						//Triggering swaps
+						fmt.Println("sender", sender.String())
+						fmt.Println("amount", amountIn.String())
+						fmt.Println("path: ", tokens)
 					}
 				}
 			}
